Remove sleep and move execution to trash method, define methods on NsxService interface
