+++ api/src/main/java/com/cloud/network/nsx/NsxNetworkRule.java
@@ -14,7 +14,7 @@
 // KIND, either express or implied.  See the License for the
 // specific language governing permissions and limitations
 // under the License.
-package org.apache.cloudstack.resource;
+package com.cloud.network.nsx;
 
 import com.cloud.network.Network;
